Refaktor. Flytt tester til servicetest for å kalle funbksjon direkte.
